# src/generated/openApi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "MeldingerReceiver API",
    "description": "Common entrypoints to interact with MeldingerReceiver.",
    "version": "v1"
  },
  "paths": {
...
    "/altinn/non-completed-instances/{appId}": {
      "get": {
        "tags": [
          "Altinn"
        ],
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AltinnMetadata"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AltinnMetadata"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AltinnMetadata"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/altinn/subscriptions/{appId}": {
      "get": {
        "tags": [
          "Altinn"
        ],
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AltinnEventsSubscription"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AltinnEventsSubscription"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AltinnEventsSubscription"
                }
              }
            }
          }
        }
      }
    },
    "/altinn/subscriptions/{id}/retrigger-altinn-validation": {
      "post": {
        "tags": [
          "Altinn"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/altinn/start-recovery-job/{appId}": {
      "post": {
        "tags": [
          "Altinn"
        ],
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecoveryJobResult"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecoveryJobResult"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecoveryJobResult"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/altinn/process/{appId}/{instanceGuid}": {
      "post": {
        "tags": [
          "Altinn"
        ],
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PostMeldingResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostMeldingResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostMeldingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/altinn/instances/{instanceGuid}/data-elements": {
      "get": {
        "tags": [
          "Altinn"
        ],
        "parameters": [
          {
            "name": "instanceGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DataElement"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DataElement"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DataElement"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/altinn/instances/{instanceGuid}/data-elements/{dataElementId}": {
      "get": {
        "tags": [
          "Altinn"
        ],
        "parameters": [
          {
            "name": "instanceGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "dataElementId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/altinn/instances/{instanceGuid}/metadata": {
      "get": {
        "tags": [
          "Altinn"
        ],
        "parameters": [
          {
            "name": "instanceGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AltinnInstance"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AltinnInstance"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AltinnInstance"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AltinnEventsSubscription": {
        "required": [
          "id",
          "callbackUrl",
          "sourceFilter",
          "createdBy",
          "consumer",
          "created",
          "validated"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "callbackUrl": {
            "type": "string"
          },
          "sourceFilter": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "consumer": {
            "type": "string"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "validated": {
            "type": "boolean"
          }
        }
      },
      "AltinnInstance": {
        "type": "object",
        "properties": {
          "id": {
            "type": [
              "null",
              "string"
            ]
          },
          "instanceOwner": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/InstanceOwner"
              }
            ]
          },
          "appId": {
            "type": [
              "null",
              "string"
            ]
          },
          "org": {
            "type": [
              "null",
              "string"
            ]
          },
          "process": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProcessState"
              }
            ]
          },
          "completeConfirmations": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/CompleteConfirmation"
            }
          },
          "data": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/DataElement"
            }
          },
          "dataValues": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "AltinnMetadata": {
        "required": [
          "instanceGuid",
          "instanceOwnerPartyId",
          "org",
          "app",
          "dataValues"
        ],
        "type": "object",
        "properties": {
          "instanceGuid": {
            "type": "string",
            "format": "uuid"
          },
          "instanceOwnerPartyId": {
            "type": "string"
          },
          "org": {
            "type": "string"
          },
          "app": {
            "type": "string"
          },
          "dataValues": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "organisationNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "processStarted": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "processEnded": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "dialogId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CompleteConfirmation": {
        "type": "object",
        "properties": {
          "stakeholderId": {
            "type": [
              "null",
              "string"
            ]
          },
          "confirmedOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "DataElement": {
        "type": "object",
        "properties": {
          "id": {
            "type": [
              "null",
              "string"
            ]
          },
          "instanceGuid": {
            "type": [
              "null",
              "string"
            ]
          },
          "dataType": {
            "type": [
              "null",
              "string"
            ]
          },
          "filename": {
            "type": [
              "null",
              "string"
            ]
          },
          "contentType": {
            "type": [
              "null",
              "string"
            ]
          },
          "size": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "contentHash": {
            "type": [
              "null",
              "string"
            ]
          },
          "isRead": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "userDefinedMetadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "fileScanResult": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FileScanResult"
              }
            ]
          }
        }
      },
...
      "FileScanResult": {
        "enum": [
          "NotApplicable",
          "Pending",
          "Clean",
          "Infected"
        ],
        "type": "string"
      },
...
      "InstanceOwner": {
        "type": "object",
        "properties": {
          "partyId": {
            "type": [
              "null",
              "string"
            ]
          },
          "personNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "organisationNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "username": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PostMeldingResponse": {
        "type": "object",
        "properties": {
          "meldingId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProcessState": {
        "type": "object",
        "properties": {
          "started": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "startEvent": {
            "type": [
              "null",
              "string"
            ]
          },
          "ended": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "endEvent": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RecoveryJobResult": {
        "required": [
          "appId"
        ],
        "type": "object",
        "properties": {
          "appId": {
            "type": "string"
          },
          "originalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "resolvedCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      }
    }
  },
...
}